requirements.txt
# Core dependencies
tor>=0.4.7
stem>=1.8.0
cryptography>=41.0.0
pycryptodome>=3.19.0
aiohttp>=3.9.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
psutil>=5.9.0
netifaces>=0.11.0

# Stealth and obfuscation
pyinstaller>=5.13.0
pyarmor>=8.3.0
obfuscator>=0.1.0

# Network security
scapy>=2.5.0
dpkt>=1.9.8
pysocks>=1.7.1

# Advanced features
selenium>=4.15.0
webdriver-manager>=4.0.0
pillow>=10.1.0
numpy>=1.26.0

# Production monitoring
prometheus-client>=0.19.0
structlog>=24.1.0
colorlog>=6.8.0